+++ src/extension.ts
@@ -2,24 +2,18 @@ import * as vscode from 'vscode';
 import * as commands from './commands';
 
 export function activate(context: vscode.ExtensionContext) {
-  const addCurrentFile = vscode.commands.registerCommand(
-    'extension.addCurrentFileToIndexDart',
-    () => {
-      commands.addCurrentFileToIndexDart();
-    },
-  );
+  const showInfoNotification = vscode.commands.registerCommand('extension.showInfo', async () => {
+    const selection = await vscode.window.showInformationMessage('Extension has been updated. Recursive export and ignore patters are now available. Check the extension settings for details.', 'Open settings');
+
+    if (selection) {
+      await vscode.commands.executeCommand('workbench.action.openSettings', 'dartBarrelExportFileGenerator');
+    }
+  });
 
   const addCurrentFileDirName = vscode.commands.registerCommand(
     'extension.addCurrentFileToIndexDartDirName',
     () => {
-      commands.addCurrentFileToIndexDartDirName();
-    },
-  );
-
-  const exportAllFiles = vscode.commands.registerCommand(
-    'extension.exportDartFilesInCurrentDirectory',
-    () => {
-      commands.exportDartFilesInCurrentDirectory();
+      commands.addCurrentFileToDartExportFile();
     },
   );
 
@@ -30,10 +24,11 @@ export function activate(context: vscode.ExtensionContext) {
     },
   );
 
-  context.subscriptions.push(addCurrentFile);
   context.subscriptions.push(addCurrentFileDirName);
-  context.subscriptions.push(exportAllFiles);
   context.subscriptions.push(exportAllFilesDirName);
+  context.subscriptions.push(showInfoNotification);
+  // show notification on start
+  vscode.commands.executeCommand('extension.showInfo');
 }
 
 export function deactivate() { }