+++ src/utils/editor-helper.ts
@@ -5,6 +5,8 @@ export const fileIsOpened = (): boolean => {
   return !!vscode.window.activeTextEditor;
 };
 
+
+/// Returns true if the current file is saved, false otherwise
 export const fileIsSaved = (): boolean => {
   if (!fileIsOpened) {
     return false;
@@ -20,6 +22,8 @@ export const fileIsSaved = (): boolean => {
   return !document.isUntitled;
 };
 
+
+/// Returns the current file path or null if the file is not opened or not saved
 export const getCurrentFilePath = (): string | null => {
   if (!(fileIsOpened() && fileIsSaved())) {
     return null;